add `buffering-indicator/ui` tests
